Restructured files for evalautor
